+++ com-mmxlabs-models-lng/com.mmxlabs.models.lng.fleet/src/com/mmxlabs/models/lng/fleet/impl/VesselStateAttributesImpl.java
@@ -46,7 +46,7 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * @generated
 	 * @ordered
 	 */
-	protected static final int NBO_RATE_EDEFAULT = 0;
+	protected static final double NBO_RATE_EDEFAULT = 0.0;
 
 	/**
 	 * The cached value of the '{@link #getNboRate() <em>Nbo Rate</em>}' attribute.
@@ -56,7 +56,7 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * @generated
 	 * @ordered
 	 */
-	protected int nboRate = NBO_RATE_EDEFAULT;
+	protected double nboRate = NBO_RATE_EDEFAULT;
 
 	/**
 	 * The default value of the '{@link #getIdleNBORate() <em>Idle NBO Rate</em>}' attribute.
@@ -66,7 +66,7 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * @generated
 	 * @ordered
 	 */
-	protected static final int IDLE_NBO_RATE_EDEFAULT = 0;
+	protected static final double IDLE_NBO_RATE_EDEFAULT = 0.0;
 
 	/**
 	 * The cached value of the '{@link #getIdleNBORate() <em>Idle NBO Rate</em>}' attribute.
@@ -76,7 +76,7 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * @generated
 	 * @ordered
 	 */
-	protected int idleNBORate = IDLE_NBO_RATE_EDEFAULT;
+	protected double idleNBORate = IDLE_NBO_RATE_EDEFAULT;
 
 	/**
 	 * The default value of the '{@link #getIdleBaseRate() <em>Idle Base Rate</em>}' attribute.
@@ -86,7 +86,7 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * @generated
 	 * @ordered
 	 */
-	protected static final int IDLE_BASE_RATE_EDEFAULT = 0;
+	protected static final double IDLE_BASE_RATE_EDEFAULT = 0.0;
 
 	/**
 	 * The cached value of the '{@link #getIdleBaseRate() <em>Idle Base Rate</em>}' attribute.
@@ -96,7 +96,7 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * @generated
 	 * @ordered
 	 */
-	protected int idleBaseRate = IDLE_BASE_RATE_EDEFAULT;
+	protected double idleBaseRate = IDLE_BASE_RATE_EDEFAULT;
 
 	/**
 	 * The default value of the '{@link #getInPortBaseRate() <em>In Port Base Rate</em>}' attribute.
@@ -106,7 +106,7 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * @generated
 	 * @ordered
 	 */
-	protected static final int IN_PORT_BASE_RATE_EDEFAULT = 0;
+	protected static final double IN_PORT_BASE_RATE_EDEFAULT = 0.0;
 
 	/**
 	 * The cached value of the '{@link #getInPortBaseRate() <em>In Port Base Rate</em>}' attribute.
@@ -116,7 +116,7 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * @generated
 	 * @ordered
 	 */
-	protected int inPortBaseRate = IN_PORT_BASE_RATE_EDEFAULT;
+	protected double inPortBaseRate = IN_PORT_BASE_RATE_EDEFAULT;
 
 	/**
 	 * The cached value of the '{@link #getFuelConsumption() <em>Fuel Consumption</em>}' containment reference list.
@@ -172,7 +172,7 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public int getNboRate() {
+	public double getNboRate() {
 		return nboRate;
 	}
 
@@ -181,8 +181,8 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setNboRate(int newNboRate) {
-		int oldNboRate = nboRate;
+	public void setNboRate(double newNboRate) {
+		double oldNboRate = nboRate;
 		nboRate = newNboRate;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET, FleetPackage.VESSEL_STATE_ATTRIBUTES__NBO_RATE, oldNboRate, nboRate));
@@ -193,7 +193,7 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public int getIdleNBORate() {
+	public double getIdleNBORate() {
 		return idleNBORate;
 	}
 
@@ -202,8 +202,8 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setIdleNBORate(int newIdleNBORate) {
-		int oldIdleNBORate = idleNBORate;
+	public void setIdleNBORate(double newIdleNBORate) {
+		double oldIdleNBORate = idleNBORate;
 		idleNBORate = newIdleNBORate;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET, FleetPackage.VESSEL_STATE_ATTRIBUTES__IDLE_NBO_RATE, oldIdleNBORate, idleNBORate));
@@ -214,7 +214,7 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public int getIdleBaseRate() {
+	public double getIdleBaseRate() {
 		return idleBaseRate;
 	}
 
@@ -223,8 +223,8 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setIdleBaseRate(int newIdleBaseRate) {
-		int oldIdleBaseRate = idleBaseRate;
+	public void setIdleBaseRate(double newIdleBaseRate) {
+		double oldIdleBaseRate = idleBaseRate;
 		idleBaseRate = newIdleBaseRate;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET, FleetPackage.VESSEL_STATE_ATTRIBUTES__IDLE_BASE_RATE, oldIdleBaseRate, idleBaseRate));
@@ -235,7 +235,7 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public int getInPortBaseRate() {
+	public double getInPortBaseRate() {
 		return inPortBaseRate;
 	}
 
@@ -244,8 +244,8 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setInPortBaseRate(int newInPortBaseRate) {
-		int oldInPortBaseRate = inPortBaseRate;
+	public void setInPortBaseRate(double newInPortBaseRate) {
+		double oldInPortBaseRate = inPortBaseRate;
 		inPortBaseRate = newInPortBaseRate;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET, FleetPackage.VESSEL_STATE_ATTRIBUTES__IN_PORT_BASE_RATE, oldInPortBaseRate, inPortBaseRate));
@@ -332,16 +332,16 @@ public class VesselStateAttributesImpl extends MMXObjectImpl implements VesselSt
 	public void eSet(int featureID, Object newValue) {
 		switch (featureID) {
 			case FleetPackage.VESSEL_STATE_ATTRIBUTES__NBO_RATE:
-				setNboRate((Integer)newValue);
+				setNboRate((Double)newValue);
 				return;
 			case FleetPackage.VESSEL_STATE_ATTRIBUTES__IDLE_NBO_RATE:
-				setIdleNBORate((Integer)newValue);
+				setIdleNBORate((Double)newValue);
 				return;
 			case FleetPackage.VESSEL_STATE_ATTRIBUTES__IDLE_BASE_RATE:
-				setIdleBaseRate((Integer)newValue);
+				setIdleBaseRate((Double)newValue);
 				return;
 			case FleetPackage.VESSEL_STATE_ATTRIBUTES__IN_PORT_BASE_RATE:
-				setInPortBaseRate((Integer)newValue);
+				setInPortBaseRate((Double)newValue);
 				return;
 			case FleetPackage.VESSEL_STATE_ATTRIBUTES__FUEL_CONSUMPTION:
 				getFuelConsumption().clear();
